# docs/swagger.yaml
basePath: /
definitions:
  service.AddAccountForm:
    properties:
      account:
        type: string
      currency:
        description: 账户计量单位可以为空
        type: string
      date:
        type: string
    required:
    - account
    - date
    type: object
  service.AddAccountTypeForm:
    properties:
      name:
        type: string
      type:
        type: string
    required:
    - name
    - type
    type: object
  service.BalanceAccountForm:
    properties:
      account:
        type: string
      date:
        type: string
      number:
        type: string
    required:
    - account
    - date
    - number
    type: object
  service.CloseAccountForm:
    properties:
      account:
        type: string
      date:
        type: string
    required:
    - account
    - date
    type: object
...
  service.SyncCommodityPriceForm:
    properties:
      commodity:
        type: string
      date:
        type: string
      price:
        type: string
    required:
    - commodity
    - date
    - price
    type: object
...
host: localhost:10000
info:
  contact:
    name: BaoXuebin
  description: Beancount-GS 是一个基于 Beancount 的个人财务管理系统，提供账本管理、交易记录、统计分析等功能。
  title: Beancount-GS API
  version: "1.0"
paths:
  /api/auth/account:
    post:
      consumes:
      - application/json
      description: 在账本中开启一个新账户
      parameters:
      - description: 新增账户表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.AddAccountForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 创建新账户
      tags:
      - 账户
  /api/auth/account/all:
    get:
      description: 返回所有未关闭账户及其币种余额和市场价値
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 获取全部账户列表（含余额）
      tags:
      - 账户
  /api/auth/account/balance:
    post:
      consumes:
      - application/json
      description: 对指定账户创建一条 balance 断言录入账本
      parameters:
      - description: 平衡账户表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.BalanceAccountForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 平衡账户余额
      tags:
      - 账户
  /api/auth/account/close:
    post:
      consumes:
      - application/json
      description: 在账本中关闭指定账户，记录关闭日期
      parameters:
      - description: 关闭账户表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CloseAccountForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 关闭账户
      tags:
      - 账户
  /api/auth/account/icon:
    post:
      consumes:
      - multipart/form-data
      description: 上传图片文件以更换指定账户的图标
      parameters:
      - description: 账户名称
        in: query
        name: account
        required: true
        type: string
      - description: 图标图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 更换账户图标
      tags:
      - 账户
  /api/auth/account/refresh:
    post:
      description: 重新加载当前账本的账户列表和货币汇率缓存
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 刷新账户缓存
      tags:
      - 账户
  /api/auth/account/type:
    get:
      description: 返回当前账本定义的所有账户类型（Key/Name 对）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 获取账户类型列表
      tags:
      - 账户
    post:
      consumes:
      - application/json
      description: 为当前账本新增一种账户类型映射
      parameters:
      - description: 新增类型表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.AddAccountTypeForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 添加账户类型
      tags:
      - 账户
  /api/auth/account/valid:
    get:
      description: 返回未关闭的所有账户，包含多货币信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 获取有效账户列表
      tags:
      - 账户
  /api/auth/commodity/currencies:
    get:
      description: 返回当前账本中所有货币及其最新汇率
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 获取所有货币列表
      tags:
      - 商品
  /api/auth/commodity/price:
    post:
      consumes:
      - application/json
      description: 向账本写入一条新的 price 条目，并刷新货币汇率缓存
      parameters:
      - description: 商品价格表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.SyncCommodityPriceForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 同步商品价格
      tags:
      - 商品
...
securityDefinitions:
  LedgerId:
    description: 需要授权的接口请在 Header 中传入 ledgerId
    in: header
    name: ledgerId
    type: apiKey
swagger: "2.0"